what i've managed to do so far
As my time has been limited, I focused on figuring out how to add and centre the 'shop' button and also how to add an ellipse.

The shop button I'm still unsure of how to achieve the rounding of the box and also centering. For the centering, I saw a few people suggesting using a flexbox could help to centre it.